try test-helper in dedicated file again, this time actually with file added
