+++ stats/test/handler.test.js
@@ -2,7 +2,8 @@ import http from 'node:http'
 import { once } from 'node:events'
 import assert from 'node:assert'
 import createDebug from 'debug'
-import { getPgPools, givenDailyParticipants } from '@filecoin-station/spark-stats-db'
+import { getPgPools } from '@filecoin-station/spark-stats-db'
+import { givenDailyParticipants } from '@filecoin-station/spark-stats-db/test-helpers.js'
 
 import { assertResponseStatus, getPort } from './test-helpers.js'
 import { createHandler } from '../lib/handler.js'
